+++ Struktury/Funkcje/det.cpp
@@ -1,6 +1,6 @@
 #include "det.h"
 
-float det(Point p1, Point p2, Point p3) {
+double det(Point p1, Point p2, Point p3) {
     // 3 x 3 
     // D = (x2 − x1)(y3 − y1) − (y2 − y1)(x3 − x1)
