+++ Struktury/Kopalnia.h
@@ -22,16 +22,16 @@ public:
 
     int getId() const;
     const Point& getLokalizacja() const;
-    float getX() const;
-    float getY() const;
+    double getX() const;
+    double getY() const;
     const std::string& getTypSurowca() const;
     int getPojemnosc() const;
     const std::vector<int>& getPrzypisaneKrasnale() const;
 
     void setId(int noweId);
     void setLokalizacja(const Point& nowaLokalizacja);
-    void setX(float noweX);
-    void setY(float noweY);
+    void setX(double noweX);
+    void setY(double noweY);
     void setTypSurowca(const std::string& nowyTypSurowca);
     void setPojemnosc(int nowaPojemnosc);
     void setPrzypisaneKrasnale(const std::vector<int>& nowePrzypisaneKrasnale);
